clean: Tree Design導入 #5
