+++ src/component/CreateAccount/index.ts
@@ -1,0 +1,1 @@
+export {CreateAccount} from './CreateAccount'
