+++ src/component/Home/element/Button/Button.stories.tsx
@@ -1,0 +1,17 @@
+import {Meta, StoryObj} from '@storybook/react'
+
+import {Button} from '.'
+
+const meta: Meta<typeof Button> = {
+  title: 'Button',
+  component: Button,
+}
+
+export default meta
+type Story = StoryObj<typeof Button>
+
+export const Primary: Story = {
+  args: {
+    sample: 'sample',
+  },
+}
