+++ src/component/Map/element/ShopInfo/element/RateModal/index.ts
