+++ src/component/Map/element/ShopInfo/element/StarRating/index.ts
