+++ src/component/Map/index.ts
@@ -1,0 +1,1 @@
+export {Map} from './Map'
