--- src/component/TopButton/index.ts
@@ -1,1 +1,0 @@
-export {TopButton} from './TopButton'
