+++ src/component/common/ResponsiveTxt/ResponsiveTxt.tsx
