+++ src/component/common/ResponsiveTxt/index.ts
