+++ src/type/info.ts
